+++ db/updates/update_11.0-11.1.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `batchIDHistory` CHANGE `id` `id` VARCHAR(255) NOT NULL;