+++ pages/manageFormula.php
@@ -475,7 +475,13 @@ if($_POST['action'] == 'deleteFormula' && $_POST['fid']){
 		$defCatClass = $settings['defCatClass'];
 		$arcID = "Archived-".$fname.$fid;
 		
-		genBatchPDF($fid,$arcID,'100','100','100',$defCatClass,$settings['qStep'],'formulas');
+		$rs = genBatchPDF($fid,$arcID,'100','100','100',$defCatClass,$settings['qStep'],$settings['defPercentage'],'formulas');
+		
+		if($rs !== true){
+			$response['error'] = 'Error archiving the formula, '.$rs['error'];
+			echo json_encode($response);
+			return;
+		}
 
 	}
 	
@@ -648,7 +654,7 @@ if($_POST['action'] == 'todo' && $_POST['fid'] && $_POST['markComplete']){
 	}
 	if(mysqli_query($conn,"UPDATE formulasMetaData SET isMade = '1', toDo = '0', madeOn = NOW(), status = '2' WHERE fid = '$fid'")){
 		$batchID = genBatchID();
-		genBatchPDF($fid,$batchID,$total_quantity,'100',$total_quantity,$defCatClass,$settings['qStep'],'makeFormula');
+		genBatchPDF($fid,$batchID,$total_quantity,'100',$total_quantity,$defCatClass,$settings['qStep'],$settings['defPercentage'],'makeFormula');
 
 		mysqli_query($conn, "DELETE FROM makeFormula WHERE fid = '$fid'");